+++ CHANGELOG.md
@@ -21,6 +21,8 @@ _Unreleased_
 
 - Rename `rye tools list` flags: `-i, --include-scripts` to `-s, --include-scripts` and `-v, --version-show` to `-v, --include-version`.  #722
 
+- Rye now adds env source to `.zprofile` for `zsh` shell.  #xxx
+
 <!-- released start -->
 
 ## 0.25.0